+++ src/components/Wowerlay.tsx
@@ -1,4 +1,16 @@
-import { PropType, Teleport, Transition, defineComponent, ref, watch } from 'vue';
+import {
+  InjectionKey,
+  PropType,
+  Teleport,
+  Transition,
+  defineComponent,
+  inject,
+  onBeforeUnmount,
+  provide,
+  reactive,
+  ref,
+  watch
+} from 'vue';
 import { WowerlayBaseProps, wowerlayBaseProps } from './WowerlayReusables';
 import { cWowerlayAnimEnter, cWowerlayAnimLeave, cWowerlayContainer } from '../consts';
 
@@ -19,6 +31,12 @@ const Emits = {
   'update:visible': (value: boolean): any => typeof value === 'boolean'
 } as const;
 
+interface ParentWowerlayContext {
+  onClose: (hook: () => void) => void;
+}
+
+const ParentWowerlayContextInjectionKey: InjectionKey<ParentWowerlayContext> = Symbol();
+
 export const Wowerlay = defineComponent({
   name: 'Wowerlay',
   inheritAttrs: false,
@@ -29,17 +47,30 @@ export const Wowerlay = defineComponent({
   emits: Emits,
   setup(props, { emit }) {
     const { onWindowClick } = useWowerlayContext();
+    const parentWowerlay = inject(ParentWowerlayContextInjectionKey, null);
+
+    const childrenWowerlayHooks = reactive([]) as Function[];
     const canClose = ref(false);
 
-    const toClass = `.${cWowerlayContainer}`;
+    const closeChildWowerlays = () => {
+      childrenWowerlayHooks.forEach((v) => v());
+    };
 
-    onWindowClick(() => {
+    const close = () => {
       if (!props.visible) return;
 
       if (canClose.value) {
+        closeChildWowerlays();
+        childrenWowerlayHooks.length = 0;
         emit('update:visible', false);
       }
-    });
+    };
+
+    parentWowerlay?.onClose(close);
+    const handleWowerlayClick = closeChildWowerlays;
+
+    onWindowClick(close);
+    onBeforeUnmount(close);
 
     watch(
       () => props.visible,
@@ -54,18 +85,30 @@ export const Wowerlay = defineComponent({
       }
     );
 
+    provide(ParentWowerlayContextInjectionKey, {
+      onClose(hook) {
+        childrenWowerlayHooks.push(hook);
+      }
+    });
+
     return {
-      toClass,
-      canClose
+      canClose,
+      handleWowerlayClick
     };
   },
   render() {
+    const toClass = `.${cWowerlayContainer}`;
+
     return (
-      <Teleport to={this.toClass}>
+      <Teleport to={toClass}>
         {/*// Todo: Add user made animation support. */}
         <Transition enterActiveClass={cWowerlayAnimEnter} leaveActiveClass={cWowerlayAnimLeave}>
           {this.visible && (
-            <WowerlayRenderer {...this.$props} {...this.$attrs}>
+            <WowerlayRenderer
+              {...this.$props}
+              {...this.$attrs}
+              onWowerlayClick={this.handleWowerlayClick}
+            >
               {this.$slots.default?.()}
             </WowerlayRenderer>
           )}